+++ package.json
@@ -62,7 +62,8 @@
     "serverless-s3-cleaner": "^2.0.1",
     "typescript": "^4.7.4",
     "vite": "^2.9.13",
-    "vitest": "^0.17.1"
+    "vitest": "^0.17.1",
+    "serverless-single-page-app-plugin": "file:./serverless-single-page-app-plugin"
   },
   "msw": {
     "workerDirectory": "public"
