+++ src/components/MainLayout/components/Header.tsx
@@ -33,7 +33,7 @@ export default function Header() {
             underline="none"
             to="/"
           >
-            My Store!
+            My Jewelry Store!
           </Link>
         </Typography>
